Переносит функции в модуль map
